requirements.txt
click>=8.0.0
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=0.19.0
rich>=12.0.0
pytest>=7.0.0
pytest-mock>=3.0.0 